date fields are now read only for jackson date fields are set to null for export (#1670)
